Updating the moveThroughMessagePages function.
To fix a bug that was force-enabling buttons.

+++ Source/ErlParser.cpp
@@ -53,7 +53,7 @@ namespace DiscordCoreInternal {
 			throw ErlParseError{ "ErlParser::readString() Error: readString() past end of buffer.\n\n" };
 		}
 		if (this->finalString.size() < this->currentSize + length) {
-			this->finalString.resize(this->finalString.size() + (length * 4));
+			this->finalString.resize(this->finalString.size() + static_cast<uint64_t>(length * 4ull));
 		}
 		const char* stringNew = static_cast<const char*>(this->dataBuffer.data()) + this->offSet;
 		this->offSet += length;
